refactor: observaciones de pull request
