+++ src/erp/data/SDataConstantsSys.java
@@ -139,7 +139,7 @@ public abstract class SDataConstantsSys {
     public static final String CFG_PARAM_SIIE_APP_URLS = "SIIE_APP_URLS";
     public static final String CFG_PARAM_SIIE_WEB_DB = "SIIE_WEB_DB";
     public static final String CFG_PARAM_SIIE_SYS_AUTH_PUSH_NOTIF = "SIIE_SYS_AUTH_PUSH_NOTIF";
-    public static final String CFG_PARAM_SWAP_SERVICE_CONFIG = "SWAP_SERVICE_CONFIG";
+    public static final String CFG_PARAM_SWAP_SERVICES_CONFIG = "SWAP_SERVICES_CONFIG";
     public static final String CFG_PARAM_SIIE_PURP_DPS_NAT = "SIIE_PURP_DPS_NAT";
     public static final String CFG_PARAM_SIIE_PURP_CFD_USE = "SIIE_PURP_CFD_USE";
     public static final String CFG_PARAM_SIIE_CFG_USR_GRP_DPS_NAT_CFD_USE = "SIIE_CFG_USR_GRP_DPS_NAT_CFD_USE";
