+++ src/erp/mod/cfg/swapss/SPublicInterface.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package erp.mod.cfg.swapms;
+package erp.mod.cfg.swapss;
 
 import erp.mod.cfg.swapms.utils.SExportUtils;
 import erp.mod.cfg.swapms.utils.SUserExport;
